remove pip related code for Gramps module
The pip related build-options was from when Gramps was not building for 5.1 in pip or setup.py in Gnome Platform 41, and is not currently necessary since setup.py works in 42 again.  It was already removed upstream, but was still in my branch and showed in in the pull request as a change.  So I removed it to get back in line with upstream for now, until it might be needed in 5.2 with pip.